Calendar moved to a component
